move captureLogs view/utils -> model/
