+++ frontend/main.js
@@ -14,6 +14,7 @@ import router from './controller/router.js'
 import { createWebSocket } from './controller/backend.js'
 import store from './model/state.js'
 import { SETTING_CURRENT_USER } from './model/database.js'
+import './model/captureLogs.js'
 import { LOGIN, LOGOUT, CONTRACT_IS_SYNCING } from './utils/events.js'
 import './utils/lazyLoadedView.js'
 import BannerGeneral from './views/components/banners/BannerGeneral.vue'
@@ -25,7 +26,6 @@ import './views/utils/translations.js'
 import './views/utils/vFocus.js'
 import './views/utils/vError.js'
 import './views/utils/vStyle.js'
-import './views/utils/captureLogs.js'
 
 import 'wicg-inert'
